+++ frontend/src/lib/components/startup/application/IncubationPlanStep.svelte
