+++ frontend/src/lib/components/startup/application/StartupDetailsStep.svelte
