+++ frontend/src/lib/components/startup/application/TeamMembersStep.svelte
